+++ src/builtins/chainables/mod.ts
@@ -1,14 +1,11 @@
 // moderate
 
-export * from "./attribute.ts";
-export * from "./convert.ts";
 export * from "./cube.ts";
 export * from "./cylinder.ts";
 export * from "./diff.ts";
 export * from "./flip.ts";
 export * from "./getBoundingBox.ts";
 export * from "./intersection.ts";
-export * from "./label.ts";
 export * from "./meld.ts";
 export * from "./mirror.ts";
 export * from "./moveTo.ts";
