+++ src/core/chainables/label.ts
@@ -1,19 +1,15 @@
-import {
-  Component,
-  ConvertibleTo,
-  Element,
-  NameHierarchy,
-  Operation,
-} from "../../core/mod.ts";
-import { Mesh } from "../Mesh.ts";
+import { Component, Element, GenericOperation } from "../chaining/mod.ts";
+import { type Hkt } from "../utils/mod.ts";
+import { NameHierarchy } from "../hierarchy/mod.ts";
+import { Product } from "../product/mod.ts";
 
 export const label = Component.create(
   "label",
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   (name: string, _show: boolean = true) =>
-    Operation.create(
+    GenericOperation.create<Product, Hkt.Identity<Product>>(
       "label",
-      async (value: Element<ConvertibleTo<Mesh>>) =>
+      async <T extends Product>(value: Element<T>) =>
         Element.applyHierarchy(
           value,
           NameHierarchy.create({
